DotCloud Build File
The DotCloud Build File, ``dotcloud.yml``, describes our stack.

We will start with a single "python" service (we will add the database later).
This service allows us to expose a WSGI-compliant web application. Django
can do WSGI out of the box, so that's perfect for us.

..

The role and syntax of the DotCloud Build File is explained in further
detail in the documentation, at http://docs.dotcloud.com/guides/build-file/.