Add PostgreSQL Database
It's time to add a database to our project! The first step is to tell
DotCloud that we want a PostgreSQL server to be added to our application.
Just edit the Build File again.

..

Note that we called our database ``db``, but it could have been anything
else, really.

If you ``dotcloud push`` again, you will see that the database service
will be created (DotCloud will notice that we added a section to the
Build File).