Django Admin Site
We will now activate the Django administration application.
Nothing is specific to DotCloud here: we just uncomment the relevant
lines of code in ``settings.py`` and ``urls.py``.

..

If we push our application now, we can go to the ``/admin`` URL,
but since we didn't call ``syncdb`` yet, the database structure
doesn't exist, and Django will refuse to do anything useful for us.