Generate the static files in ~/volatile instead of ~/data
With the new "Granite" builder, DotCloud applications are built in the
background while your production application is still running. When the
build is finished the old service is stopped and its "~/data" directory
is migrated to the new service.

This change in the way applications are built means that the "~/data"
directory is only accessible during the initial push of an application.
In subsequent pushes the "~/data" directory is migrated when the build
is over which means that the build can no longer modify the "~/data"
directory.

This commit simply changes the path where static files are generated
(STATIC_ROOT) to the "~/volatile/static" directory.